refactor(lotto): relocate files to proper packages and delete neeless files
