+++ src/main/kotlin/lotto/domain/Lottos.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class Lottos(private val lottos: List<Lotto>) {
