+++ src/main/kotlin/lotto/view/InputView.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.view
 
 object InputView {
     fun readPayment(): Int {
