+++ src/test/kotlin/lotto/domain/LottosTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
